+++ CHANGES.rst
@@ -27,6 +27,12 @@ Lemaitre <glemaitre>`, :user:`Adrin Jalali <adrinjalali>`, :user:`Olivier Grisel
 <ogrisel>` and others participated through many discussions in defining the
 requirements and the public API.
 
+- Added the `AdaptiveSquashingTransformer` that
+robustly rescales and smoothly clips numerical columns,
+enabling more robust handling of numerical columns
+with neural networks.
+  :pr:`1310` by :user:`David Holzmüller <dholzmueller>`
+
 Changes
 -------